Refactor updateField method in UsuarioController
This commit refactors the `updateField` method in the `UsuarioController` class. The code now checks if the field being updated is the password field and applies bcrypt encryption if it is. The updated value is then assigned to the corresponding field in the `$usuario` object. This change improves the security of the application.