Refactor indexCliente method in ListaUniformeController
This commit refactors the indexCliente method in the ListaUniformeController by replacing a single line of code. The line that retrieves a single record from the database using the find method is replaced with a query that retrieves multiple records using the where method. This change improves the efficiency and readability of the code.